Fix unintentional `list`, `search` and `info` deprecations
By no longer inheriting from `QueryCommand`.
