+++ Manifest.txt
@@ -400,6 +400,7 @@ lib/rubygems/path_support.rb
 lib/rubygems/platform.rb
 lib/rubygems/psych_additions.rb
 lib/rubygems/psych_tree.rb
+lib/rubygems/query_utils.rb
 lib/rubygems/rdoc.rb
 lib/rubygems/remote_fetcher.rb
 lib/rubygems/request.rb
