+++ lib/rubygems/commands/info_command.rb
@@ -1,14 +1,19 @@
 # frozen_string_literal: true
 
 require 'rubygems/command'
-require 'rubygems/commands/query_command'
+require 'rubygems/query_utils'
 
-class Gem::Commands::InfoCommand < Gem::Commands::QueryCommand
+class Gem::Commands::InfoCommand < Gem::Command
+
+  include Gem::QueryUtils
 
   def initialize
-    super "info", "Show information for the given gem"
+    super "info", "Show information for the given gem",
+         :name => //, :domain => :local, :details => false, :versions => true,
+         :installed => nil, :version => Gem::Requirement.default
 
-    remove_option('--name-matches')
+    add_query_options
+
     remove_option('-d')
 
     defaults[:details] = true
