+++ lib/rubygems/commands/search_command.rb
@@ -1,15 +1,17 @@
 # frozen_string_literal: true
 require 'rubygems/command'
-require 'rubygems/commands/query_command'
+require 'rubygems/query_utils'
 
-class Gem::Commands::SearchCommand < Gem::Commands::QueryCommand
+class Gem::Commands::SearchCommand < Gem::Command
+
+  include Gem::QueryUtils
 
   def initialize
-    super 'search', 'Display remote gems whose name matches REGEXP'
+    super 'search', 'Display remote gems whose name matches REGEXP',
+         :name => //, :domain => :remote, :details => false, :versions => true,
+         :installed => nil, :version => Gem::Requirement.default
 
-    remove_option '--name-matches'
-
-    defaults[:domain] = :remote
+    add_query_options
   end
 
   def arguments # :nodoc:
