+++ test/rubygems/test_gem_gem_runner.rb
@@ -76,4 +76,34 @@ class TestGemGemRunner < Gem::TestCase
     assert_equal "WARNING:  query command is deprecated. It will be removed on or after 2019-12-01.\n", @ui.error
   end
 
+  def test_info_succeeds
+    args = %w[info]
+
+    use_ui @ui do
+      assert_nil @runner.run(args)
+    end
+
+    assert_empty @ui.error
+  end
+
+  def test_list_succeeds
+    args = %w[list]
+
+    use_ui @ui do
+      assert_nil @runner.run(args)
+    end
+
+    assert_empty @ui.error
+  end
+
+  def test_search_succeeds
+    args = %w[search]
+
+    use_ui @ui do
+      assert_nil @runner.run(args)
+    end
+
+    assert_empty @ui.error
+  end
+
 end
